Updated Orderable trait (split Pivot behavior into complementary trait)
